Undo refactoring of DBDefinitionFactory
